pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.31",
    "langchain-elasticsearch>=0.3.2",
    "langchain-openai>=0.3.35",
    "python-dotenv>=1.1.1",
]
